Renamed cow-subgraph to subgraph
