+++ src/index.ts
@@ -3,4 +3,4 @@ export { CowError } from './utils/common'
 export { ALL_SUPPORTED_CHAIN_IDS, SupportedChainId } from './constants/chains'
 export { COW_PROTOCOL_SETTLEMENT_CONTRACT_ADDRESS } from './constants'
 export * from './types'
-export * as GraphQL from './cow-subgraph/graphql'
+export * as GraphQL from './subgraph/graphql'
